More changes found in my working directory
At some point (around 17 October 2023 according to the modification
dates of the files) I abandoned work on this, and today I found these
changes in my home directory. I recall that I had proceeded to the point
where jibri was working when it was in the same server as the rest, but
I had trouble making a jibri running on another server communicate to
the prosody and jvb containers.

In the end I decided to abandon docker and instead I created the
grnet.jitsi Ansible collection that can be found at
https://github.com/grnet/team-collaboration-toolkit.

The changes in this and the previous two commits are

© 2023 GRNET (https://grnet.gr/)